organized the files and packages a bit
