+++ internal/endpoints/routes.go
@@ -1,4 +1,4 @@
-package authfox
+package endpoints
 
 import (
 	"github.com/gin-gonic/gin"
